refactor: move assets folder
