+++ assets/products.ts
